Updating/Editing App to call Components
