+++ src/Components/Attendees.js
@@ -1,3 +1,5 @@
+import { useState } from "react";
 export default function Attendees() {
+  const [showAttendees, setShowAttendees] = useState(false);
   return;
 }
