+++ src/Components/Event.js
@@ -1,3 +1,11 @@
-export default function Event() {
+import { useState } from "react";
+export default function Event({ event }) {
+  const [events, setEvents] = useState(eventsData);
+  
+  function handleSubmit(e) {
+    e.preventDefault();
+    addEvent();
+    resetEventForm();
+  }
   return;
 }
